[WIP] Release 0.6.0 with TF 2.4.0 support.

+++ modules/jni/src/main/scala/org/platanios/tensorflow/jni/TensorFlow.scala
@@ -15,18 +15,20 @@
 
 package org.platanios.tensorflow.jni
 
-import com.typesafe.scalalogging.Logger
-import org.slf4j.LoggerFactory
+import scala.io.{Codec, Source}
 
 import java.io.{IOException, InputStream}
-import java.nio.file.{Files, Path, StandardCopyOption}
+import java.nio.charset.StandardCharsets
+import java.nio.file.{Files, Path, Paths, StandardCopyOption}
 
-import scala.io.{Codec, Source}
+import com.typesafe.scalalogging.Logger
+import org.slf4j.LoggerFactory
 
 /**
   * @author Emmanouil Antonios Platanios
   */
 object TensorFlow {
+
   private val logger: Logger = Logger(LoggerFactory.getLogger("TensorFlow Native"))
 
   /** TensorFlow native library name. */
@@ -41,12 +43,20 @@ object TensorFlow {
   /** TensorFlow ops library name. */
   private val OPS_LIB_NAME: String = "tensorflow_ops"
 
+  /** Directory in the user home directory that stores TensorFlow Scala meta-data. For example, paths to the extracted
+    * TensorFlow binaries are stored there in order to avoid extracting them every time TensorFlow Scala loads.
+    */
+  private val TF_SCALA_DIRECTORY: Path = Paths.get(System.getProperty("user.home"), ".tf-scala")
+
+  /** Name of the file that holds information about the extracted TensorFlow native libraries. */
+  private val EXTRACTED_LIBS_FILENAME: String = "extracted-libs.txt"
+
   /** Current platform. */
   private val platform = {
     val name = System.getProperty("os.name").toLowerCase
     if (name.contains("linux")) "linux"
     else if (name.contains("os x") || name.contains("darwin")) "darwin"
-    else if (name.contains("windows")) "windows-cpu"
+    else if (name.contains("windows")) "windows"
     else name.replaceAll("\\s", "")
   }
 
@@ -59,50 +69,21 @@ object TensorFlow {
     //     loaded (for example, tensorflow/examples/android and tensorflow/contrib/android include the required native
     //     code in differently named libraries).
     // then it seems that the native library has already been loaded and there is nothing else to do.
-    if (!checkIfLoaded()) {
+    if (!checkIfLoaded() && !checkIfExtracted()) {
       // Native code is not present, perhaps it has been packaged into the JAR file containing this code.
       val tempDirectory = Files.createTempDirectory("tensorflow_scala_native_libraries")
-      tempDirectory.toFile.deleteOnExit()
-      val classLoader = Thread.currentThread.getContextClassLoader
+      // tempDirectory.toFile.deleteOnExit()
+      loadLibrariesFromDirectory(tempDirectory)
 
-      // Check if a TensorFlow native framework library resources are provided and load them.
-      (makeResourceNames(LIB_FRAMEWORK_NAME) ++ makeResourceNames(LIB_NAME)).foreach {
-        case (name, path, preLoad) =>
-          val resource = extractResource(name, classLoader.getResourceAsStream(path), tempDirectory)
-          if (preLoad) {
-            resource.foreach(r => System.load(r.toAbsolutePath.toString))
-          }
+      // Store the temporary directory path in the TF Scala meta-data directory, if it is accessible.
+      if (Files.exists(TF_SCALA_DIRECTORY) || TF_SCALA_DIRECTORY.toFile.mkdirs()) {
+        val extractedLibsFile = TF_SCALA_DIRECTORY.resolve(EXTRACTED_LIBS_FILENAME)
+        Files.write(extractedLibsFile, tempDirectory.toAbsolutePath.toString.getBytes(StandardCharsets.UTF_8))
       }
-
-      // Load the TensorFlow JNI bindings from the appropriate resource.
-      val jniPaths = makeResourceNames(JNI_LIB_NAME).flatMap {
-        case (name, path, _) => extractResource(name, classLoader.getResourceAsStream(path), tempDirectory)
-      }
-      if (jniPaths.isEmpty) {
-        throw new UnsatisfiedLinkError(
-          s"Cannot find the TensorFlow JNI bindings for platform: $platform. See " +
-              "https://github.com/eaplatanios/tensorflow_scala/tree/master/README.md for possible solutions " +
-              "(such as building the library from source).")
-      }
-      jniPaths.foreach(path => {
-        try {
-          System.load(path.toAbsolutePath.toString)
-        } catch {
-          case exception: IOException => throw new UnsatisfiedLinkError(
-            "Unable to load the TensorFlow JNI bindings from the extracted file. This could be due to the TensorFlow " +
-                s"native library not being available. Error: ${exception.getMessage}.")
-        }
-      })
-
-      // Load the TensorFlow ops library from the appropriate resource.
-      val opsPaths = makeResourceNames(OPS_LIB_NAME).flatMap {
-        case (name, path, _) => extractResource(name, classLoader.getResourceAsStream(path), tempDirectory)
-      }
-      opsPaths.foreach(path => loadOpLibrary(path.toAbsolutePath.toString))
     }
   }
 
-  /** Checks if the TensorFlow JNI bindings library has been loaded. */
+  /** Checks if the TensorFlow native libraries have been loaded. */
   private def checkIfLoaded(): Boolean = {
     try {
       TensorFlow.version
@@ -112,10 +93,72 @@ object TensorFlow {
     }
   }
 
+  /** Checks if the TensorFlow native libraries have been previously extracted, and loads them if they have been. */
+  private def checkIfExtracted(): Boolean = {
+    val extractedLibsFile = TF_SCALA_DIRECTORY.resolve(EXTRACTED_LIBS_FILENAME)
+    if (Files.exists(extractedLibsFile)) {
+      try {
+        val source    = Source.fromFile(extractedLibsFile.toFile)
+        val directory = Paths.get(source.getLines.mkString)
+        source.close()
+        loadLibrariesFromDirectory(directory)
+      } catch {
+        case _: Throwable => false
+      }
+    } else {
+      false
+    }
+  }
+
+  /** Attempts to load the TensorFlow native libraries from the provided [[Path]] and returns `true` if they were
+    * loaded successfully.
+    */
+  private def loadLibrariesFromDirectory(directory: Path): Boolean = {
+    directory.toFile.deleteOnExit()
+    val classLoader = Thread.currentThread.getContextClassLoader
+
+    // Check if a TensorFlow native framework library resources are provided and load them.
+    (makeResourceNames(LIB_FRAMEWORK_NAME) ++ makeResourceNames(LIB_NAME)).foreach {
+      case (name, path, preLoad) =>
+        val resource = extractResource(name, classLoader.getResourceAsStream(path), directory)
+        if (preLoad) {
+          resource.foreach(r => System.load(r.toAbsolutePath.toString))
+        }
+    }
+
+    // Load the TensorFlow JNI bindings from the appropriate resource.
+    val jniPaths = makeResourceNames(JNI_LIB_NAME).flatMap {
+      case (name, path, _) => extractResource(name, classLoader.getResourceAsStream(path), directory)
+    }
+    if (jniPaths.isEmpty) {
+      throw new UnsatisfiedLinkError(
+        s"Cannot find the TensorFlow JNI bindings for platform: $platform. See " +
+            "https://github.com/eaplatanios/tensorflow_scala/tree/master/README.md for possible solutions " +
+            "(such as building the library from source).")
+    }
+    jniPaths.foreach(path => {
+      try {
+        System.load(path.toAbsolutePath.toString)
+      } catch {
+        case exception: IOException => throw new UnsatisfiedLinkError(
+          "Unable to load the TensorFlow JNI bindings from the extracted file. This could be due to the TensorFlow " +
+              s"native library not being available. Error: ${exception.getMessage}.")
+      }
+    })
+
+    // Load the TensorFlow ops library from the appropriate resource.
+    val opsPaths = makeResourceNames(OPS_LIB_NAME).flatMap {
+      case (name, path, _) => extractResource(name, classLoader.getResourceAsStream(path), directory)
+    }
+    opsPaths.foreach(path => loadOpLibrary(path.toAbsolutePath.toString))
+
+    checkIfLoaded()
+  }
+
   /** Maps the provided library name to a set of filenames, similar to [[System.mapLibraryName]], but considering all
     * combinations of `dylib` and `so` extensions, along with versioning for TensorFlow 2.x. */
   private def mapLibraryName(lib: String): Seq[(String, Boolean)] = {
-    if (platform == "windows-cpu") {
+    if (platform == "windows") {
       Seq((s"$lib.dll", true))
     } else if (lib == JNI_LIB_NAME || lib == OPS_LIB_NAME) {
       Seq((s"lib$lib.so", false))
@@ -123,10 +166,10 @@ object TensorFlow {
       Seq(
         (s"lib$lib.so.link", false),
         (s"lib$lib.so.2.link", false),
-        (s"lib$lib.so.2.3.1", false),
+        (s"lib$lib.so.2.4.0", false),
         (s"lib$lib.dylib.link", false),
         (s"lib$lib.2.dylib.link", false),
-        (s"lib$lib.2.3.1.dylib", false),
+        (s"lib$lib.2.4.0.dylib", false),
       )
     }
   }
@@ -145,22 +188,24 @@ object TensorFlow {
   private def extractResource(filename: String, resourceStream: InputStream, directory: Path): Option[Path] = {
     if (resourceStream != null) {
       val filePath = directory.resolve(filename)
-      logger.debug(s"Extracting the '$filename' native library to ${filePath.toAbsolutePath}.")
-      try {
-        val (streamToCopy, path) = if (filename.endsWith(".link")) {
-          val linkPathSource = Source.fromInputStream(resourceStream)(Codec.UTF8)
-          val linkPath = linkPathSource.mkString
-          val stream = Thread.currentThread.getContextClassLoader.getResourceAsStream(linkPath)
-          linkPathSource.close()
-          (stream, directory.resolve(filename.dropRight(5)))
-        } else {
-          (resourceStream, filePath)
+      if (!Files.exists(filePath)) {
+        logger.debug(s"Extracting the '$filename' native library to ${filePath.toAbsolutePath}.")
+        try {
+          val (streamToCopy, path) = if (filename.endsWith(".link")) {
+            val linkPathSource = Source.fromInputStream(resourceStream)(Codec.UTF8)
+            val linkPath       = linkPathSource.mkString
+            val stream         = Thread.currentThread.getContextClassLoader.getResourceAsStream(linkPath)
+            linkPathSource.close()
+            (stream, directory.resolve(filename.dropRight(5)))
+          } else {
+            (resourceStream, filePath)
+          }
+          val numBytes             = Files.copy(streamToCopy, path, StandardCopyOption.REPLACE_EXISTING)
+          logger.debug(String.format(s"Copied $numBytes bytes to ${filePath.toAbsolutePath}."))
+        } catch {
+          case exception: Exception =>
+            throw new UnsatisfiedLinkError(s"Error while extracting the '$filename' native library: $exception.")
         }
-        val numBytes = Files.copy(streamToCopy, path, StandardCopyOption.REPLACE_EXISTING)
-        logger.debug(String.format(s"Copied $numBytes bytes to ${filePath.toAbsolutePath}."))
-      } catch {
-        case exception: Exception =>
-          throw new UnsatisfiedLinkError(s"Error while extracting the '$filename' native library: $exception")
       }
       Some(filePath)
     } else {
@@ -178,6 +223,9 @@ object TensorFlow {
   @native def loadOpLibrary(libraryPath: String): Array[Byte]
   @native def enableXLA(): Unit
 
+  @native def setLogLevel(value: String): Unit
+  @native def getLogLevel: String
+
   //region Internal API
 
   @native private[tensorflow] def updateInput(
@@ -185,8 +233,13 @@ object TensorFlow {
   @native private[tensorflow] def addControlInput(graphHandle: Long, opHandle: Long, inputOpHandle: Long): Int
   @native private[tensorflow] def clearControlInputs(graphHandle: Long, opHandle: Long): Int
   @native private[tensorflow] def setRequestedDevice(graphHandle: Long, opHandle: Long, device: String): Int
+
   @native private[tensorflow] def setAttributeProto(
-      graphHandle: Long, opHandle: Long, attributeName: String, attributeValue: Array[Byte]): Unit
+      graphHandle: Long,
+      opHandle: Long,
+      attributeName: String,
+      attributeValue: Array[Byte],
+  ): Unit
 
   //endregion Internal API
 }
